+++ src/index.tsx
@@ -35,7 +35,7 @@ export { default as Stars } from 'stars'
 export { default as Stepper } from 'stepper'
 export { default as SuccessModal } from 'successModal'
 export { default as Tabs } from 'tabs'
-export { default as Text } from 'text'
+export { default as Text, TextDisplayType, TextTagType } from 'text'
 export { default as TextField } from 'textField'
 export { default as TimePicker } from 'timePicker'
 export { default as Title } from 'title'
